+++ changelog.md
@@ -6,6 +6,7 @@ More info frames is no longer clamped to screen.
 
 - *Selecting an item for "Award Later" during a session will now properly count the item as awarded.*
 - *Tier token items are now properly registered as such for usage in more info tooltips.*
+- *BoE's are recognized once again.*
 
 # 3.13.1